NetBSD Lintified.

+++ strings.c
@@ -1,4 +1,4 @@
-/* $Id: strings.c,v 1.17 2009/02/24 13:46:54 kristaps Exp $ */
+/* $Id: strings.c,v 1.18 2009/02/24 14:52:55 kristaps Exp $ */
 /*
  * Copyright (c) 2008 Kristaps Dzonsons <kristaps@kth.se>
  *
@@ -62,9 +62,9 @@ mdoc_isescape(const char *p)
 	case ('e'):
 		return(2);
 	case ('('):
-		if (0 == *++p || ! isgraph(*p))
+		if (0 == *++p || ! isgraph((int)*p))
 			return(0);
-		if (0 == *++p || ! isgraph(*p))
+		if (0 == *++p || ! isgraph((int)*p))
 			return(0);
 		return(4);
 	case ('['):
@@ -74,7 +74,7 @@ mdoc_isescape(const char *p)
 	}
 
 	for (c = 3, p++; *p && ']' != *p; p++, c++)
-		if ( ! isgraph(*p))
+		if ( ! isgraph((int)*p))
 			break;
 
 	return(*p == ']' ? c : 0);
